turn back into float (in case of np weirdness)

+++ dataClasses.py
@@ -512,7 +512,7 @@ def tieFor2Estimate(probs):
 
     #print(unnormalized)
     normFactor = sum(unnormalized)
-    result = [u/normFactor for u in unnormalized]
+    result = [float(u/normFactor) for u in unnormalized]
     print(result)
     return result
